labs/Day1_Advanced/requirements.txt
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-experimental>=0.1.0
langchain-text-splitters>=0.3.0
faiss-cpu>=1.8.0
pypdf>=5.0.0
python-dotenv>=1.0.0
duckduckgo-search>=6.0.0
tiktoken>=0.8.0

